# contracts/openapi.yaml
openapi: 3.1.1
info:
  title: Shadowcook API
  version: 0.1.0
paths:
...
  /admin/authentication-settings:
    get:
      operationId: getAuthenticationSettings
      security: [{ sessionCookie: [] }]
      responses:
        {
          '200':
            {
              description: Settings,
              content:
                {
                  application/json:
                    {
                      schema:
                        {
                          type: object,
                          required: [loginMode],
                          properties:
                            {
                              loginMode:
                                {
                                  type: string,
                                  enum: [PASSWORD_ONLY, EMAIL_CODE_ONLY, PASSWORD_OR_EMAIL_CODE],
                                },
                            },
                        },
                    },
                },
            },
        }
    put:
      operationId: saveAuthenticationSettings
      security: [{ sessionCookie: [] }]
      requestBody:
        required: true
        content:
          {
            application/json:
              {
                schema:
                  {
                    type: object,
                    required: [loginMode],
                    properties:
                      {
                        loginMode:
                          {
                            type: string,
                            enum: [PASSWORD_ONLY, EMAIL_CODE_ONLY, PASSWORD_OR_EMAIL_CODE],
                          },
                      },
                  },
              },
          }
      responses:
        {
          '204': { description: Saved. },
          '409':
            {
              description: SMTP required for email-only mode,
              content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } },
            },
        }
  /admin/ingredients:
    get:
      operationId: listIngredients
      summary: Lists instance-owned ingredients with aliases and recipe usage counts.
      security: [{ sessionCookie: [] }]
      responses:
        {
          '200':
            {
              description: Ingredients,
              content:
                {
                  application/json:
                    {
                      schema:
                        {
                          type: object,
                          required: [ingredients],
                          properties:
                            {
                              ingredients:
                                { type: array, items: { $ref: '#/components/schemas/Ingredient' } },
                            },
                        },
                    },
                },
            },
        }
    post:
      operationId: createIngredient
      security: [{ sessionCookie: [] }]
      requestBody:
        {
          required: true,
          content:
            { application/json: { schema: { $ref: '#/components/schemas/IngredientInput' } } },
        }
      responses:
        {
          '201':
            {
              description: Ingredient created,
              content:
                { application/json: { schema: { $ref: '#/components/schemas/Ingredient' } } },
            },
          '400':
            {
              description: Invalid ingredient,
              content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } },
            },
          '409':
            {
              description: Ingredient conflict,
              content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } },
            },
        }
  /admin/ingredients/{publicId}:
    patch:
      operationId: updateIngredient
      security: [{ sessionCookie: [] }]
      parameters:
        [{ name: publicId, in: path, required: true, schema: { type: string, format: uuid } }]
      requestBody:
        {
          required: true,
          content:
            { application/json: { schema: { $ref: '#/components/schemas/IngredientInput' } } },
        }
      responses:
        {
          '204': { description: Ingredient updated. },
          '400':
            {
              description: Invalid ingredient,
              content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } },
            },
          '404':
            {
              description: Ingredient not found,
              content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } },
            },
        }
    delete:
      operationId: deleteIngredient
      security: [{ sessionCookie: [] }]
      parameters:
        [{ name: publicId, in: path, required: true, schema: { type: string, format: uuid } }]
      responses:
        {
          '204': { description: Ingredient deleted. },
          '404':
            {
              description: Ingredient not found,
              content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } },
            },
          '409':
            {
              description: Ingredient is in use,
              content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } },
            },
        }
  /admin/ingredients/{publicId}/aliases:
    post:
      operationId: createIngredientAlias
      security: [{ sessionCookie: [] }]
      parameters:
        [{ name: publicId, in: path, required: true, schema: { type: string, format: uuid } }]
      requestBody:
        {
          required: true,
          content:
            { application/json: { schema: { $ref: '#/components/schemas/IngredientAliasInput' } } },
        }
      responses:
        {
          '201':
            {
              description: Alias created,
              content:
                { application/json: { schema: { $ref: '#/components/schemas/IngredientAlias' } } },
            },
          '400':
            {
              description: Invalid alias,
              content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } },
            },
          '404':
            {
              description: Ingredient not found,
              content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } },
            },
          '409':
            {
              description: Alias conflict,
              content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } },
            },
        }
  /admin/ingredients/{publicId}/aliases/{aliasPublicId}:
    patch:
      operationId: updateIngredientAlias
      security: [{ sessionCookie: [] }]
      parameters:
        [
          { name: publicId, in: path, required: true, schema: { type: string, format: uuid } },
          { name: aliasPublicId, in: path, required: true, schema: { type: string, format: uuid } },
        ]
      requestBody:
        {
          required: true,
          content:
            { application/json: { schema: { $ref: '#/components/schemas/IngredientAliasInput' } } },
        }
      responses:
        {
          '204': { description: Alias updated. },
          '400':
            {
              description: Invalid alias,
              content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } },
            },
          '404':
            {
              description: Alias not found,
              content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } },
            },
          '409':
            {
              description: Alias conflict,
              content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } },
            },
        }
    delete:
      operationId: deleteIngredientAlias
      security: [{ sessionCookie: [] }]
      parameters:
        [
          { name: publicId, in: path, required: true, schema: { type: string, format: uuid } },
          { name: aliasPublicId, in: path, required: true, schema: { type: string, format: uuid } },
        ]
      responses:
        {
          '204': { description: Alias deleted. },
          '404':
            {
              description: Alias not found,
              content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } },
            },
        }
...
components:
...
  schemas:
...
    IngredientInput:
      type: object
      additionalProperties: false
      required: [canonicalName]
      properties: { canonicalName: { type: string, minLength: 1, maxLength: 160 } }
    IngredientAliasInput:
      type: object
      additionalProperties: false
      required: [alias]
      properties: { alias: { type: string, minLength: 1, maxLength: 160 } }
    IngredientAlias:
      type: object
      additionalProperties: false
      required: [publicId, alias, localizationKey]
      properties:
        publicId: { type: string, format: uuid }
        alias: { type: string }
        localizationKey: { type: [string, 'null'] }
    Ingredient:
      type: object
      additionalProperties: false
      required: [publicId, canonicalName, localizationKey, usageCount, aliases]
      properties:
        publicId: { type: string, format: uuid }
        canonicalName: { type: string }
        localizationKey: { type: [string, 'null'] }
        usageCount: { type: integer, minimum: 0 }
        aliases: { type: array, items: { $ref: '#/components/schemas/IngredientAlias' } }
...
    ApiError:
      type: object
      additionalProperties: false
      required: [code, error]
      properties:
        code:
          type: string
          enum:
            - AUTHENTICATION_FAILED
            - AUTHENTICATION_REQUIRED
            - INVALID_CREDENTIALS
            - INVALID_LOGIN_BODY
            - INVALID_PASSWORD
            - PASSWORD_CHANGE_REQUIRED
            - RECIPE_NOT_FOUND
            - RECIPE_CONFLICT
            - INVALID_RECIPE
            - INVALID_RECIPE_CATEGORY
            - PUBLISH_CATEGORY_REQUIRED
            - RECIPE_NOT_PUBLISHED
            - INVALID_UNIT
            - UNIT_CONFLICT
            - UNIT_NOT_FOUND
            - UNIT_IN_USE
            - INVALID_CONVERSION
            - UNIT_DIMENSION_MISMATCH
            - INVALID_INGREDIENT
            - INGREDIENT_CONFLICT
            - INGREDIENT_NOT_FOUND
            - INGREDIENT_IN_USE
            - INVALID_ALIAS
            - ALIAS_CONFLICT
            - ALIAS_NOT_FOUND
        error:
          type: string
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: shadowcook_session